# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  github_com_osi-oss_osi_internal_dto.ErrorResponse:
    description: Стандартный формат ответа при ошибке
    properties:
      error:
        description: |-
          Сообщение об ошибке
          Example: invalid email or password
        example: invalid email or password
        type: string
    type: object
...
  github_com_osi-oss_osi_internal_dto.LoginRequest:
    description: Учётные данные для входа в систему
    properties:
      email:
        description: |-
          Email пользователя
          Example: user@example.com
        example: user@example.com
        type: string
      password:
        description: |-
          Пароль пользователя
          Example: SecurePass123
        example: SecurePass123
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  github_com_osi-oss_osi_internal_dto.LoginResponse:
    description: Ответ при успешной авторизации с JWT токеном
    properties:
      message:
        description: |-
          Сообщение о результате
          Example: Login successful
        example: Login successful
        type: string
      token:
        description: |-
          JWT токен для авторизации запросов
          Example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ
        type: string
    type: object
  github_com_osi-oss_osi_internal_dto.MessageResponse:
    description: Стандартный формат успешного ответа с сообщением
    properties:
      message:
        description: |-
          Сообщение о результате операции
          Example: operation completed successfully
        example: operation completed successfully
        type: string
    type: object
...
  github_com_osi-oss_osi_internal_dto.PasswordResetRequest:
    description: Email для отправки инструкций по сбросу пароля
    properties:
      email:
        description: |-
          Email пользователя для восстановления доступа
          Example: user@example.com
        example: user@example.com
        type: string
    required:
    - email
    type: object
...
  github_com_osi-oss_osi_internal_dto.ProfileResponse:
    description: Ответ с данными профиля пользователя
    properties:
      message:
        description: |-
          Сообщение о результате
          Example: Profile retrieved successfully
        example: Profile retrieved successfully
        type: string
      user:
        allOf:
        - $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.UserResponse'
        description: Данные пользователя
    type: object
  github_com_osi-oss_osi_internal_dto.ResetPasswordRequest:
    description: Токен сброса и новый пароль
    properties:
      new_password:
        description: |-
          Новый пароль (минимум 6 символов)
          Example: NewSecurePass456
        example: NewSecurePass456
        minLength: 6
        type: string
      token:
        description: |-
          Токен из письма для сброса пароля
          Example: a1b2c3d4e5f6g7h8i9j0
        example: a1b2c3d4e5f6g7h8i9j0
        type: string
    required:
    - new_password
    - token
    type: object
  github_com_osi-oss_osi_internal_dto.SignUpRequest:
    description: Данные для регистрации нового пользователя в системе
    properties:
      email:
        description: |-
          Email пользователя (должен быть уникальным)
          Example: user@example.com
        example: user@example.com
        type: string
      password:
        description: |-
          Пароль (минимум 6 символов)
          Example: SecurePass123
        example: SecurePass123
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  github_com_osi-oss_osi_internal_dto.SignUpResponse:
    description: Ответ при успешной регистрации пользователя
    properties:
      message:
        description: |-
          Сообщение о результате
          Example: User created successfully
        example: User created successfully
        type: string
      user:
        allOf:
        - $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.UserResponse'
        description: Данные созданного пользователя
    type: object
  github_com_osi-oss_osi_internal_dto.TokenValidationResponse:
    description: Результат проверки токена сброса пароля
    properties:
      message:
        description: |-
          Сообщение о результате
          Example: Token is valid
        example: Token is valid
        type: string
      valid:
        description: |-
          Валиден ли токен
          Example: true
        example: true
        type: boolean
    type: object
...
  github_com_osi-oss_osi_internal_dto.UserResponse:
    description: Публичные данные пользователя
    properties:
      created_at:
        description: |-
          Дата регистрации
          Example: 2024-01-15T10:30:00Z
        example: "2024-01-15T10:30:00Z"
        type: string
      email:
        description: |-
          Email пользователя
          Example: user@example.com
        example: user@example.com
        type: string
      email_verified:
        description: |-
          Подтверждён ли email
          Example: true
        example: true
        type: boolean
      first_name:
        description: |-
          Имя пользователя
          Example: Иван
        example: Иван
        type: string
      id:
        description: |-
          Уникальный идентификатор пользователя
          Example: 1
        example: 1
        type: integer
      last_name:
        description: |-
          Фамилия пользователя
          Example: Петров
        example: Петров
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API для управления организациями, локациями, отделами и позициями
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: OSI API
  version: "1.0"
paths:
  /forgot-password:
    post:
      consumes:
      - application/json
      description: |-
        Отправляет письмо с инструкциями по сбросу пароля на указанный email.
        В целях безопасности всегда возвращает успех, даже если email не найден.
        Токен сброса действителен 1 час.
      parameters:
      - description: Email для восстановления доступа
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.PasswordResetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Инструкции отправлены (если email существует)
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.MessageResponse'
        "400":
          description: Неверный формат email
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.ErrorResponse'
      summary: Запрос сброса пароля
      tags:
      - Аутентификация
  /login:
    post:
      consumes:
      - application/json
      description: |-
        Авторизует пользователя по email и паролю.
        Возвращает JWT токен, который нужно передавать в заголовке Authorization: Bearer <token>.
        Токен действителен 24 часа.
      parameters:
      - description: Учётные данные для входа
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успешная авторизация, возвращается JWT токен
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.LoginResponse'
        "400":
          description: 'Ошибка валидации: неверный формат запроса'
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.ErrorResponse'
        "401":
          description: Неверный email или пароль
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.ErrorResponse'
      summary: Авторизация пользователя
      tags:
      - Аутентификация
  /logout:
    post:
      consumes:
      - application/json
      description: |-
        Завершает сессию пользователя.
        Удаляет токен из cookies (auth_token).
        После выхода JWT токен остаётся валидным до истечения срока.
      produces:
      - application/json
      responses:
        "200":
          description: Успешный выход из системы
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.MessageResponse'
      security:
      - BearerAuth: []
      summary: Выход из системы
      tags:
      - Аутентификация
...
  /profile:
    get:
      consumes:
      - application/json
      description: |-
        Возвращает данные профиля авторизованного пользователя.
        Требует авторизацию через JWT токен.
      produces:
      - application/json
      responses:
        "200":
          description: Данные профиля пользователя
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.ProfileResponse'
        "401":
          description: Отсутствует или невалидный токен авторизации
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.ErrorResponse'
        "404":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получение профиля текущего пользователя
      tags:
      - Пользователь
  /reset-password:
    post:
      consumes:
      - application/json
      description: |-
        Устанавливает новый пароль с использованием токена из письма.
        Токен можно использовать только один раз.
        Новый пароль должен быть не менее 6 символов.
      parameters:
      - description: Токен сброса и новый пароль
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Пароль успешно изменён
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.MessageResponse'
        "400":
          description: Невалидный или истёкший токен, или пароль слишком короткий
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.ErrorResponse'
      summary: Сброс пароля
      tags:
      - Аутентификация
  /reset-password/validate:
    get:
      consumes:
      - application/json
      description: |-
        Проверяет, действителен ли токен для сброса пароля.
        Используется для предварительной проверки перед отображением формы сброса.
      parameters:
      - description: Токен из письма для сброса пароля
        example: abc123def456
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Токен валиден, можно сбрасывать пароль
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.TokenValidationResponse'
        "400":
          description: Токен не указан, невалиден или истёк
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.ErrorResponse'
      summary: Проверка токена сброса пароля
      tags:
      - Аутентификация
  /signup:
    post:
      consumes:
      - application/json
      description: |-
        Создаёт нового пользователя в системе.
        Email должен быть уникальным, пароль минимум 6 символов.
        После успешной регистрации пользователь может авторизоваться через /login.
      parameters:
      - description: Данные для регистрации
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.SignUpRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Пользователь успешно создан
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.SignUpResponse'
        "400":
          description: 'Ошибка валидации: неверный формат email или пароль короче
            6 символов'
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.ErrorResponse'
        "409":
          description: Пользователь с таким email уже существует
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.ErrorResponse'
      summary: Регистрация нового пользователя
      tags:
      - Аутентификация
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"